requirements.txt
torchvision
magickwand
pillow
opencv-python
opencv-contrib-python
scikit-image
numpy
Wand
